refactor lib/packages.js
