+++ routes/packages.js
@@ -1,10 +1,31 @@
 'use strict'
 
 const r = require('koa-router')()
+const path = require('path')
+const url = require('url')
 const packages = require('../lib/packages')
 const tarballs = require('../lib/tarballs')
 const config = require('../config')
-const path = require('path')
+
+function addShaToPath (p, sha) {
+  let ext = path.extname(p)
+  let filename = path.basename(p, ext)
+  p = path.dirname(p)
+
+  p = path.join(p, `${filename}/${sha}${ext}`)
+  return p
+}
+
+function rewriteTarballURLs (pkg, host, protocol) {
+  for (let version of Object.keys(pkg.versions)) {
+    let dist = pkg.versions[version].dist
+    let u = url.parse(dist.tarball)
+    u.pathname = addShaToPath(u.pathname, dist.shasum)
+    u.host = host
+    u.protocol = protocol
+    dist.tarball = url.format(u)
+  }
+}
 
 // get package metadata
 r.get('/:name', function * () {
@@ -20,7 +41,7 @@ r.get('/:name', function * () {
     return
   }
   let cloudfront = this.headers['user-agent'] === 'Amazon CloudFront'
-  packages.rewriteTarballURLs(pkg, cloudfront ? config.cloudfrontHost : this.headers.host, this.request.protocol)
+  rewriteTarballURLs(pkg, cloudfront ? config.cloudfrontHost : this.headers.host, this.request.protocol)
   this.set('ETag', pkg.etag)
   this.set('Cache-Control', `public, max-age=${config.cache.packageTTL}`)
   this.body = pkg
